+++ api/src/main/java/com/example/market/api/dtos/ProductDto.java
@@ -1,12 +1,14 @@
 package com.example.market.api.dtos;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
     private String category;
 
-    public ProductDto(Long id, String title, int price, String category) {
+    public ProductDto(Long id, String title, BigDecimal price, String category) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -31,11 +33,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
